Add tests and small fixes for CandyGuards
CandyGuards are currently untested. This commit adds basic tests for
creating a candy machine with limit guards. Initialize guard accounts are
also tested and small fixes to make them pass are added. The wrap
instruction is also added to be able to apply guards to machines.
