+++ example/CandyMachine/config_lines/1.tres
@@ -1,5 +1,5 @@
 [gd_resource type="ConfigLine" format=3 uid="uid://sesikyny42c8"]
 
 [resource]
-name = "test_nft1"
-uri = "test_uri1"
+name = "test_nft2"
+uri = "test_uri2"
